Fix Unmarshal behavior around interface{}s and maps
When unmarshalling into an interface{}, we were trying to use v.Set() on
the interface's concrete value. This doesn't work. Revamp our setters to
track the interface and the concrete value separately.

When unmarshalling a dictionary into a map, we weren't handling the
values properly. You can't call .Set() on the result of .MapIndex().
Instead of trying to preserve (and update) the existing value, which
even encoding/json doesn't try, just create a new value each time.