Unmarshal CFNumbers into interface{}s correctly
When unmarshalling a CFNumber into an interface{}, we were just cramming
it into an int64. This didn't work for floating point values. We could
try cramming it into a float64, but that won't work for larger integer
values. Instead, teach Unmarshal how to construct the correct numeric
type for a given CFNumber.